+++ src/classes/Route.ts
@@ -4,7 +4,7 @@ import parseURL from "@/functions/parseURL"
 import { EndFn, RateLimitConfig, RealAny } from "@/types/internal"
 import { isRegExp } from "util/types"
 import { UsableValidator } from "@/classes/Validator"
-import { PathItemObject } from "openapi3-ts/dist/oas31"
+import { PathItemObject } from "openapi3-ts/oas31"
 
 import HttpRequestContext from "@/classes/request/HttpRequestContext"
 import WsOpenContext from "@/classes/request/WsOpenContext"