+++ src/types/global.ts
@@ -19,4 +19,35 @@ export type URLEncodedParsed = Record<string, string>
 
 export type ParsedBody = JSONParsed | URLEncodedParsed | string
 
+export type RatelimitInfos = {
+	/**
+	 * The Number of hits the client made in the current time window
+	 * @since 8.6.0
+	*/ hits: number
+	/**
+	 * The Maximum number of hits the client is allowed to make in the specified time window
+	 * @since 8.6.0
+	*/ maxHits: number
+	/**
+	 * Whether the client has recieved the penalty
+	 * @since 8.6.0
+	*/ hasPenalty: boolean
+	/**
+	 * The Number of milliseconds penalty the client recieves
+	 * @since 8.6.0
+	*/ penalty: number
+	/**
+	 * The Number of milliseconds a time window is long
+	 * @since 8.6.0
+	*/ timeWindow: number
+	/**
+	 * The Date when the time Window (+ penalty) is over
+	 * @since 8.6.0
+	*/ endsAt: Date
+	/**
+	 * The Number of milliseconds until the time Window (+ penalty) is over
+	 * @since 8.6.0
+	*/ endsIn: number
+}
+
 export { type Content, ParseStream } from "@/functions/parseContent"