+++ templates/[Javascript] Basic JSON API/src/routes/echo.js
@@ -2,7 +2,7 @@ import { fileRouter } from "../index.js"
 
 export default new fileRouter.Path('/')
   .http('POST', '/', (http) => http
-    .onRequest((ctr) => {
-      ctr.print({ content: ctr.rawBody, requests: ctr["@"].requests })
+    .onRequest(async(ctr) => {
+      ctr.print({ content: await ctr.rawBody('utf8'), requests: ctr["@"].requests })
     })
   )